区域发展/requirements.txt
pypdf==5.1.0
python-dotenv==1.0.0
requests==2.31.0
